requirements.txt
django>=2.0
django-debug-toolbar
gunicorn
psycopg2
whitenoise
python-telegram-bot
lxml